requirements.txt
Django==4.0.1
djangorestframework==3.13.1
python-decouple==3.3
djangorestframework-simplejwt==5.0.0

pre-commit==2.17.0
